Expand application model and create form

+++ backend/src/services/applications.service.ts
@@ -1,25 +1,70 @@
-import { Application } from '../types/application';
+import { Application } from '../@types/application';
 import { pool } from '../database/database';
 
 
-export async function getAllApplications(): Promise<Application[]> {
-  const result = await pool.query('SELECT * FROM applications');
-
+export async function getAllApplications(userId: number): Promise<Application[]> {
+  const result = await pool.query(
+      `
+      SELECT *
+      FROM applications
+      WHERE user_id = $1
+      ORDER BY id DESC;
+      `,
+      [userId]
+    );
   return result.rows;
+}
+
+export async function getApplicationById(id: number, userId: number) {
+  const result = await pool.query(
+    `
+    SELECT *
+    FROM applications
+    WHERE id = $1
+      AND user_id = $2
+    `,
+    [id, userId]
+  );
+
+  return result.rows[0];
 }
 
 export async function createApplication(
   company: string,
-  position: string
+  position: string,
+  status: string,
+  location: string,
+  salary: string,
+  notes: string,
+  jobUrl: string,
+  userId: number
 ): Promise<Application> {
   const result = await pool.query(
     `
     INSERT INTO applications
-    (company, position, status)
-    VALUES ($1, $2, $3)
+    (
+      company,
+      position,
+      status,
+      location,
+      salary,
+      notes,
+      job_url,
+      user_id
+    )
+    VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
     RETURNING *;
     `,
-    [company, position, 'Applied']
+    [
+      company,
+      position,
+      status,
+      location,
+      salary,
+      notes,
+      jobUrl,
+      userId,
+    ]
   );
 
   return result.rows[0];
@@ -30,6 +75,7 @@ export async function deleteApplication(id: number): Promise<boolean> {
     `
     DELETE FROM applications
     WHERE id = $1
+    AND user_id = $2
     RETURNING *;
     `,
     [id]
@@ -49,6 +95,7 @@ export async function updateApplication(
     SET company = $1,
         position = $2
     WHERE id = $3
+    AND user_id = $4
     RETURNING *;
     `,
     [company, position, id]
@@ -57,6 +104,29 @@ export async function updateApplication(
   if (result.rowCount === 0) {
     return null;
   }
+
+  return result.rows[0];
+}
+
+export async function uploadResume(
+  applicationId: number,
+  userId: number,
+  filename: string
+) {
+  const result = await pool.query(
+    `
+    UPDATE applications
+    SET resume_path = $1
+    WHERE id = $2
+      AND user_id = $3
+    RETURNING *;
+    `,
+    [
+      filename,
+      applicationId,
+      userId,
+    ]
+  );
 
   return result.rows[0];
 }
